fix: remove static response function | create json response
