+++ api/controllers/gcp_billing.go
@@ -1,13 +1,14 @@
 package controllers
 
 import (
+	"net/http"
+
 	"boilerplate-api/external_services/gcp"
 	"boilerplate-api/internal/api_errors"
-	"boilerplate-api/internal/api_response"
 	"boilerplate-api/internal/config"
+	"boilerplate-api/internal/json_response"
 	"cloud.google.com/go/billing/budgets/apiv1/budgetspb"
 	"google.golang.org/api/cloudbilling/v1"
-	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
@@ -39,10 +40,10 @@ func (cc GCPBillingController) GetCost(c *gin.Context) {
 		cc.logger.Error("Error fetching Billing Info records", err.Error())
 		err := api_errors.InternalError.Wrap(err, "Failed To Find Billing info for GCP project")
 		status, errM := api_errors.HandleError(err)
-		c.JSON(status, api_response.Error{Error: errM})
+		c.JSON(status, json_response.Error{Error: errM})
 		return
 	}
-	c.JSON(http.StatusOK, api_response.Data[*cloudbilling.ProjectBillingInfo]{
+	c.JSON(http.StatusOK, json_response.Data[*cloudbilling.ProjectBillingInfo]{
 		Data: billingData,
 	})
 }
@@ -54,10 +55,10 @@ func (cc GCPBillingController) GetBudgetInfo(c *gin.Context) {
 		cc.logger.Error("Error fetching Billing Info records", err.Error())
 		err := api_errors.InternalError.Wrap(err, "Failed To Find Billing info for GCP project")
 		status, errM := api_errors.HandleError(err)
-		c.JSON(status, api_response.Error{Error: errM})
+		c.JSON(status, json_response.Error{Error: errM})
 		return
 	}
-	c.JSON(http.StatusOK, api_response.Data[*budgetspb.Budget]{
+	c.JSON(http.StatusOK, json_response.Data[*budgetspb.Budget]{
 		Data: billingData,
 	})
 }
@@ -69,10 +70,10 @@ func (cc GCPBillingController) CreateUpdateBudget(c *gin.Context) {
 		cc.logger.Error("Error fetching Billing Info records", err.Error())
 		err := api_errors.InternalError.Wrap(err, "Failed To Find Billing info for GCP project")
 		status, errM := api_errors.HandleError(err)
-		c.JSON(status, api_response.Error{Error: errM})
+		c.JSON(status, json_response.Error{Error: errM})
 		return
 	}
-	c.JSON(http.StatusOK, api_response.Data[*budgetspb.Budget]{
+	c.JSON(http.StatusOK, json_response.Data[*budgetspb.Budget]{
 		Data: billingData,
 	})
 }
