+++ internal/middlewares/jwt.go
@@ -2,10 +2,10 @@ package middlewares
 
 import (
 	"boilerplate-api/internal/api_errors"
-	"boilerplate-api/internal/api_response"
 	"boilerplate-api/internal/auth"
 	"boilerplate-api/internal/config"
 	"boilerplate-api/internal/constants"
+	"boilerplate-api/internal/json_response"
 
 	"github.com/getsentry/sentry-go"
 	"github.com/gin-gonic/gin"
@@ -42,7 +42,7 @@ func (m JWTAuthMiddleWare) Handle() gin.HandlerFunc {
 			m.logger.Error("Error getting token from header: ", err.Error())
 			err = api_errors.Unauthorized.Wrap(err, "Error getting token from header")
 			status, errM := api_errors.HandleError(err)
-			c.JSON(status, api_response.Error{Error: errM})
+			c.JSON(status, json_response.Error{Error: errM})
 			c.Abort()
 			return
 		}
@@ -52,7 +52,7 @@ func (m JWTAuthMiddleWare) Handle() gin.HandlerFunc {
 			m.logger.Error("Error parsing token: ", parseErr.Error())
 			err = api_errors.Unauthorized.Wrap(parseErr, "Failed to parse and verify token")
 			status, errM := api_errors.HandleError(err)
-			c.JSON(status, api_response.Error{Error: errM})
+			c.JSON(status, json_response.Error{Error: errM})
 			c.Abort()
 			return
 		}
@@ -62,7 +62,7 @@ func (m JWTAuthMiddleWare) Handle() gin.HandlerFunc {
 			m.logger.Error("Error retrieving claims: ", claimsError.Error())
 			err = api_errors.Unauthorized.Wrap(claimsError, "Failed to retrieve claims from token")
 			status, errM := api_errors.HandleError(err)
-			c.JSON(status, api_response.Error{Error: errM})
+			c.JSON(status, json_response.Error{Error: errM})
 			c.Abort()
 			return
 		}
