+++ internal/types/base.go
@@ -1,4 +1,4 @@
-package models
+package types
 
 import (
 	"time"
@@ -6,10 +6,14 @@ import (
 	"gorm.io/gorm"
 )
 
-// Base contains common columns for all tables.
-type Base struct {
-	ID        int64          `json:"id"`
+type TimeStamps struct {
 	CreatedAt time.Time      `json:"created_at"`
 	UpdatedAt time.Time      `json:"updated_at"`
 	DeletedAt gorm.DeletedAt `gorm:"index" json:"deleted_at"` //add soft delete in gorm
+}
+
+// BaseModal contains common columns for all tables.
+type BaseModal struct {
+	ID int64 `json:"id"`
+	TimeStamps
 }
