--- models/email.go
@@ -1,15 +1,0 @@
-package models
-
-type EmailParams struct {
-	To              string
-	SubjectData     string
-	SubjectTemplate string
-	BodyData        interface{}
-	BodyTemplate    string
-	Lang            string
-}
-
-type EmailSubject struct {
-	Data string
-}
-
